try to fix the unit tests
