+++ app.js
@@ -1,5 +1,4 @@
 const express = require('express');
-// const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const bookRouter = require('./routes/bookRouter')();
 require('dotenv').config();
