+++ models/dbConfig.js
@@ -1,19 +1,12 @@
-// const dbConfig = {
-//   type: 'postgres',
-//   host: process.env.DATABASE_HOST,
-//   port: 5432,
-//   username: process.env.DATABASE_USERNAME,
-//   password: process.env.DATABASE_PASSWORD,
-//   database: process.env.DATABASE_NAME,
-// };
+require('dotenv').config();
 
 const dbConfig = {
   type: 'postgres',
-  host: 'localhost',
+  host: process.env.DATABASE_HOST,
   port: 5432,
-  username: 'postgres',
-  password: '12345',
-  database: 'node-library',
+  username: process.env.DATABASE_USERNAME,
+  password: process.env.DATABASE_PASSWORD,
+  database: process.env.DATABASE_NAME,
 };
 
 module.exports = dbConfig;
