+++ tests/mockBooks.js
@@ -1,0 +1,33 @@
+const mockBooks = [{
+  id: 1,
+  title: 'Claire DeWitt and the City of the Dead',
+  author: 'Sara Gran',
+  genre: 'Crime',
+  read: false
+}, {
+  id: 2,
+  title: 'Gone Girl',
+  author: 'Gillian Flynn',
+  genre: 'Crime',
+  read: true,
+}, {
+  id: 3,
+  title: 'The Thief',
+  author: 'Fuminori Nakamura',
+  genre: 'Crime',
+  read: false
+}, {
+  id: 4,
+  title: 'American Dirt',
+  author: 'Jeanine Cummins',
+  genre: 'Crime',
+  read: false
+}, {
+  id: 5,
+  title: 'The Time Machine',
+  author: 'H. G. Wells',
+  genre: 'Science Fiction',
+  read: false
+}];
+
+module.exports = mockBooks;
